feature(feedback-card): add comment count
