+++ src/components/feedback/CommentCard/CommentCard.tsx
@@ -8,6 +8,7 @@ import { useParams } from "next/navigation";
 import { DestructiveAction } from "components/core";
 import {
   useDeleteCommentMutation,
+  useFeedbackByIdQuery,
   useInfiniteCommentsQuery,
 } from "generated/graphql";
 import { app } from "lib/config";
@@ -64,11 +65,17 @@ const CommentCard = ({
   const { mutate: deleteComment, isPending: isDeletePending } =
     useDeleteCommentMutation({
       onSettled: () =>
-        queryClient.invalidateQueries({
-          queryKey: useInfiniteCommentsQuery.getKey({
-            feedbackId,
+        Promise.all([
+          queryClient.invalidateQueries({
+            queryKey: useInfiniteCommentsQuery.getKey({
+              feedbackId,
+            }),
           }),
-        }),
+          queryClient.invalidateQueries({ queryKey: ["Posts.infinite"] }),
+          queryClient.invalidateQueries({
+            queryKey: useFeedbackByIdQuery.getKey({ rowId: feedbackId }),
+          }),
+        ]),
     });
 
   const actionIsPending = isPending || isDeletePending;
