+++ src/components/feedback/CreateComment/CreateComment.tsx
@@ -9,6 +9,7 @@ import { z } from "zod";
 import { CharacterLimit } from "components/core";
 import {
   useCreateCommentMutation,
+  useFeedbackByIdQuery,
   useInfiniteCommentsQuery,
 } from "generated/graphql";
 import { app } from "lib/config";
@@ -47,11 +48,17 @@ const CreateComment = () => {
     onSettled: () => {
       reset();
 
-      return queryClient.invalidateQueries({
-        queryKey: useInfiniteCommentsQuery.getKey({
-          feedbackId,
+      return Promise.all([
+        queryClient.invalidateQueries({
+          queryKey: useInfiniteCommentsQuery.getKey({
+            feedbackId,
+          }),
         }),
-      });
+        queryClient.invalidateQueries({ queryKey: ["Posts.infinite"] }),
+        queryClient.invalidateQueries({
+          queryKey: useFeedbackByIdQuery.getKey({ rowId: feedbackId }),
+        }),
+      ]);
     },
   });
