+++ src/components/feedback/VotingButtons/VotingButtons.tsx
@@ -1,10 +1,11 @@
-import { HStack, Icon, Tooltip } from "@omnidev/sigil";
+import { HStack, Icon, Text, Tooltip } from "@omnidev/sigil";
 import {
   PiArrowFatLineDown,
   PiArrowFatLineDownFill,
   PiArrowFatLineUp,
   PiArrowFatLineUpFill,
 } from "react-icons/pi";
+import { match } from "ts-pattern";
 
 import { app } from "lib/config";
 import {
@@ -12,20 +13,7 @@ import {
   useHandleUpvoteMutation,
 } from "lib/hooks/mutations";
 
-import type { TooltipTriggerProps, VstackProps } from "@omnidev/sigil";
 import type { Downvote, Post, Project, Upvote } from "generated/graphql";
-import type { IconType } from "react-icons";
-
-interface VoteButtonProps extends TooltipTriggerProps {
-  /** Number of votes (upvotes or downvotes). */
-  votes: number | undefined;
-  /** Tooltip text. */
-  tooltip: string;
-  /** Visual icon. */
-  icon: IconType;
-  /** Props to pass to the main content container. */
-  contentProps?: VstackProps;
-}
 
 interface Props {
   /** Feedback ID. */
@@ -73,58 +61,78 @@ const VotingButtons = ({
 
   const isVotePending = isUpvotePending || isDownvotePending;
 
-  const VOTE_BUTTONS: VoteButtonProps[] = [
-    {
-      id: "upvote",
-      votes: totalUpvotes,
-      tooltip: app.feedbackPage.details.upvote,
-      icon: upvote ? PiArrowFatLineUpFill : PiArrowFatLineUp,
-      color: "brand.tertiary",
-      onClick: (e) => {
-        e.stopPropagation();
-        handleUpvote();
-      },
-      disabled: isVotePending || isOptimistic,
-    },
-    {
-      id: "downvote",
-      votes: totalDownvotes,
-      tooltip: app.feedbackPage.details.downvote,
-      icon: downvote ? PiArrowFatLineDownFill : PiArrowFatLineDown,
-      color: "brand.quinary",
-      onClick: (e) => {
-        e.stopPropagation();
-        handleDownvote();
-      },
-      disabled: isVotePending || isOptimistic,
-    },
-  ];
+  const netTotalVotes = totalUpvotes - totalDownvotes;
+
+  const netVotesColor = match(netTotalVotes)
+    .with(0, () => "gray.400")
+    .when(
+      (net) => net > 0,
+      () => "brand.tertiary",
+    )
+    .otherwise(() => "brand.quinary");
+
+  const netVotesSign = match(netTotalVotes)
+    .when(
+      (net) => net > 0,
+      () => "+",
+    )
+    .otherwise(() => "");
 
   return (
-    <HStack
-      position="absolute"
-      top={{ base: 1.5, sm: 3.5 }}
-      right={{ base: 4, sm: 6 }}
-    >
-      {VOTE_BUTTONS.map(({ id, votes, tooltip, icon, ...rest }) => (
-        <Tooltip
-          key={id}
-          hasArrow={false}
-          trigger={
-            <HStack gap={2} py={1} fontVariant="tabular-nums">
-              <Icon src={icon} w={5} h={5} />
-              {votes}
-            </HStack>
-          }
-          triggerProps={{
-            variant: "icon",
-            bgColor: "transparent",
-            ...rest,
-          }}
-        >
-          {tooltip}
-        </Tooltip>
-      ))}
+    <HStack gap={1} justify="center" placeSelf="flex-start" mr={-2.5} mt={-2}>
+      <Tooltip
+        hasArrow={false}
+        trigger={
+          <Icon
+            src={upvote ? PiArrowFatLineUpFill : PiArrowFatLineUp}
+            w={5}
+            h={5}
+          />
+        }
+        triggerProps={{
+          variant: "icon",
+          bgColor: "transparent",
+          color: "brand.tertiary",
+          onClick: (e) => {
+            e.stopPropagation();
+            handleUpvote();
+          },
+          disabled: isVotePending || isOptimistic,
+        }}
+      >
+        {app.feedbackPage.details.upvote}
+      </Tooltip>
+
+      <Text
+        color={netVotesColor}
+        whiteSpace="nowrap"
+        fontVariant="tabular-nums"
+      >
+        {`${netVotesSign}${netTotalVotes}`}
+      </Text>
+
+      <Tooltip
+        hasArrow={false}
+        trigger={
+          <Icon
+            src={downvote ? PiArrowFatLineDownFill : PiArrowFatLineDown}
+            w={5}
+            h={5}
+          />
+        }
+        triggerProps={{
+          variant: "icon",
+          bgColor: "transparent",
+          color: "brand.quinary",
+          onClick: (e) => {
+            e.stopPropagation();
+            handleDownvote();
+          },
+          disabled: isVotePending || isOptimistic,
+        }}
+      >
+        {app.feedbackPage.details.downvote}
+      </Tooltip>
     </HStack>
   );
 };
